# Tweak SepParserAvx512PackCmpOrMoveMaskTzcnt by moving MoveMask